Refactor: 컴포넌트 트리 개선
